refactor(cli): extract EDIT cluster into src/cli/issue/edit.rs (Seam B)
Behavior-preserving shard per the 2026-06-25 architecture analysis (refactor-2026-06-25/, DO-PARTIAL verdict, Seam B). Moves the entire issue-edit cluster (handle_edit + bulk field/label handlers + build_labels_edited_fields + render_bulk_edit_results + project_key_from_issue_key + type-error classifier/hint consts + JQL_CONFIRM_THRESHOLD + 4 inline test blocks) verbatim out of create.rs (2,447→394 LOC) into a new edit.rs (2,067 LOC). create.rs now well under the ADR-0012 threshold. No logic change. All behavioral invariants preserved byte-for-byte: --label single-vs-bulk endpoint fork, --type camelCase/lowercase asymmetry, cross-project --type exit-64 guard, --field+--label mutual-exclusion exit-64, --type 400 dual-gate classifier, #398 echo asymmetry. Two include_str!("create.rs")→("edit.rs") retargets + guard-comment update for the label-conflict meta-tests. Test counts unchanged (1957 passed / 93 ignored before AND after — verified). parse_field_kv stays pub(crate) in create.rs (shared with edit.rs + jsm_create.rs). ADR-0012.
